# [SofaHelper] Reduce AdvancedTimer overhead